Enable grade level to take multiple digit number
The program does not allow user to take multiple digit number for grade
level.

However, some music examination board can take more than a decade to
finish in the worst case. Thus, it is important to allow the program to
take multiple digit number for grade level for the sake of flexibility.

Allow the program to take multiple digit number for grade level.

Modified Regex expression to allow multiple digit number for grade level
and added test cases to test the new requirements.